requirements.txt
python-telegram-bot[webhooks]==21.6
python-dotenv==1.0.1
aiosqlite==0.20.0
httpx==0.28.1
flask==3.0.3
